feat(guard): policy bundle compiler and local parser (HGC061-HGC075)
